+++ packages/keyring-eth-ledger-bridge/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump `@metamask/hw-wallet-sdk` from `^0.8.0` to `^0.9.0` ([#576](https://github.com/MetaMask/accounts/pull/576))
 - Bump `@metamask/keyring-api` from `^23.3.0` to `^23.4.0` ([#583](https://github.com/MetaMask/accounts/pull/583))
 
+### Fixed
+
+- Throw `HardwareWalletError` with `DeviceStateOnlyV4Supported` when `signTypedData` is called with a non-V4 version ([#575](https://github.com/MetaMask/accounts/pull/575))
+
 ## [12.2.0]
 
 ### Added